# errors: add a benchmark comparing stack trace performance