Cost is roughly 1600 ns to 6000 ns depending on the stack size. Adds two additional allocations.

BenchmarkErrors/pkg/errors-stack-10-4         	 1000000	      1676 ns/op	     320 B/op	       3 allocs/op
BenchmarkErrors/errors-stack-10-4             	20000000	        64.9 ns/op	      16 B/op	       1 allocs/op
BenchmarkErrors/pkg/errors-stack-100-4        	  500000	      3058 ns/op	     320 B/op	       3 allocs/op
BenchmarkErrors/errors-stack-100-4            	 3000000	       472 ns/op	      16 B/op	       1 allocs/op
BenchmarkErrors/pkg/errors-stack-1000-4       	  200000	     10685 ns/op	     320 B/op	       3 allocs/op
BenchmarkErrors/errors-stack-1000-4           	  300000	      4523 ns/op	      16 B/op	       1 allocs/op

Fixes #72